Added POST request DTOs for quizzes and leaderboards; Support for SwaggerUI API docs; Bug fixes

+++ backend/src/main/java/com/backend/WhoSaidIt/DTOs/leaderboard/TimeAttackEntryDTO.java
@@ -4,6 +4,6 @@ public record TimeAttackEntryDTO(
         Long id,
         String playerName,
         Integer score,
-        Integer timeTaken,
-        Integer averageTimePerQuestion
+        Double timeTaken,
+        Double averageTimePerQuestion
 ) implements LeaderboardEntryDTO {}
